backend/requirements.txt
flask==3.0.3
google-cloud-speech
google-cloud-language
google-generativeai